[FIX] Convert int to NSInteger to avoid any problems in 64 bits archs

+++ src/tests/FBTestBlocker.m
@@ -24,8 +24,8 @@
 @end
 
 @implementation FBTestBlocker {
-    int _signalsRemaining;
-    int _expectedSignalCount;
+    NSInteger _signalsRemaining;
+    NSInteger _expectedSignalCount;
 }
 
 - (id)init {
